BRS-4215 Delete User only if exists

+++ controllers/pguser_controller.go
@@ -191,9 +191,18 @@ func (r *PgUserReconciler) createPgApi(ctx context.Context, user *apiV1.PgUser) 
 func (r *PgUserReconciler) finalize(ctx context.Context, user *apiV1.PgUser, pgApi pgapi.PgRoleAPI) error {
 	logger := log.FromContext(ctx)
 
-	if err := pgApi.DeleteRole(user.Name); err != nil {
-		logger.Error(err, "Unable to remove login role "+user.Name+" from "+user.GetInstanceIdString())
+	// Delete only if user exists
+	exists, err := pgApi.IsRoleExisting(user.Name)
+	if err != nil {
+		logger.Error(err, "Unable to check users existence "+user.Name+" from "+user.GetInstanceIdString())
 		return err
+	}
+
+	if exists {
+		if err := pgApi.DeleteRole(user.Name); err != nil {
+			logger.Error(err, "Unable to remove login role "+user.Name+" from "+user.GetInstanceIdString())
+			return err
+		}
 	}
 
 	// Update Login Role Exists Condition
@@ -202,21 +211,22 @@ func (r *PgUserReconciler) finalize(ctx context.Context, user *apiV1.PgUser, pgA
 		return err
 	}
 
-	// Delete Secret
-	roleSecret := coreV1.Secret{
-		ObjectMeta: metaV1.ObjectMeta{
-			Namespace: user.Namespace,
-			Name:      user.Spec.Secret.Name,
-		},
-	}
-	if err := r.Delete(ctx, &roleSecret); err != nil {
-		logger.Error(err, "Unable to delete Secret")
+	// Delete Secret if exists
+	roleSecret := coreV1.Secret{}
+	exists, err = getResource(ctx, r, types.NamespacedName{Namespace: user.Namespace, Name: user.Spec.Secret.Name}, &roleSecret)
+	if err != nil {
 		return err
+	}
+	if exists {
+		if err := r.Delete(ctx, &roleSecret); err != nil {
+			logger.Error(err, "Unable to delete Secret")
+			return err
+		}
 	}
 
 	// Remove finalizer
 	controllerutil.RemoveFinalizer(user, apiV1.DefaultFinalizerPgUser)
-	err := r.Update(ctx, user)
+	err = r.Update(ctx, user)
 	if err != nil {
 		return err
 	}
